#6 rename models to model_checkpoints and add example checkpoints
