+++ bacpipe/pipelines/birdnet.py
@@ -16,7 +16,9 @@ class Model(ModelBaseClass):
 
     def __init__(self):
         super().__init__(sr=SAMPLE_RATE, segment_length=LENGTH_IN_SAMPLES)
-        model = tf.keras.models.load_model("bacpipe/models/birdnet", compile=False)
+        model = tf.keras.models.load_model(
+            self.MODEL_BASE_PATH + "/birdnet", compile=False
+        )
         self.model = tf.keras.Sequential(model.embeddings_model)
 
     def preprocess(self, audio):
