+++ bacpipe/pipelines/hbdet.py
@@ -98,7 +98,7 @@ class Model(ModelBaseClass):
     def __init__(self):
         super().__init__(sr=SAMPLE_RATE, segment_length=LENGTH_IN_SAMPLES)
         orig_model = tf.keras.models.load_model(
-            "bacpipe/models/hbdet",
+            self.MODEL_BASE_PATH + "/hbdet",
             custom_objects={"FBetaScote": metrics.FBetaScore},
         )
         model_list = orig_model.layers[:-2]
