+++ bacpipe/pipelines/rcl_fs_bsed.py
@@ -19,7 +19,8 @@ class Model(ModelBaseClass):
         super().__init__(sr=SAMPLE_RATE, segment_length=LENGTH_IN_SAMPLES)
         self.model = ResNet()
         state_dict = torch.load(
-            "bacpipe/models/rcl_fs_bsed/bioacoustics_model.pth", weights_only=True
+            self.MODEL_BASE_PATH + "/rcl_fs_bsed/bioacoustics_model.pth",
+            weights_only=True,
         )
         enc_sd = state_dict["encoder"]
         drop_keys = ["lin.0.weight", "lin.0.bias", "lin.2.weight", "lin.2.bias"]
